Fix unit tests with new objectify
